sanity test changes explorer docs
